Move on_show_status_pane_change()
