+++ foo_ui_columns/fcl_layout.cpp
@@ -1,6 +1,7 @@
 #include "stdafx.h"
 #include "layout.h"
 #include "font_notify.h"
+#include "main_window.h"
 #include "rebar.h"
 
 class export_layout : public cui::fcl::dataset {
