+++ foo_ui_columns/font_notify.h
@@ -20,7 +20,6 @@ extern HFONT g_plist_font;
 
 void on_show_toolbars_change();
 void on_show_status_change();
-void on_show_status_pane_change();
 void on_status_font_change();
 
 void font_cleanup();
